+++ cassandra/connection.py
@@ -1,8 +1,15 @@
 import errno
 from functools import wraps, partial
 import logging
+import sys
 from threading import Event, RLock
-from Queue import Queue
+import time
+import traceback
+
+if 'gevent.monkey' in sys.modules:
+    from gevent.queue import Queue
+else:
+    from Queue import Queue  # noqa
 
 from cassandra import ConsistencyLevel, AuthenticationFailed, OperationTimedOut
 from cassandra.marshal import int8_unpack, int32_pack
@@ -150,16 +157,99 @@ class Connection(object):
         raise NotImplementedError()
 
     def defunct(self, exc):
-        raise NotImplementedError()
+        with self.lock:
+            if self.is_defunct or self.is_closed:
+                return
+            self.is_defunct = True
 
-    def send_msg(self, msg, cb):
-        raise NotImplementedError()
+        trace = traceback.format_exc(exc)
+        if trace != "None":
+            log.debug("Defuncting connection (%s) to %s: %s\n%s",
+                      id(self), self.host, exc, traceback.format_exc(exc))
+        else:
+            log.debug("Defuncting connection (%s) to %s: %s", id(self), self.host, exc)
 
-    def wait_for_response(self, msg, **kwargs):
-        raise NotImplementedError()
+        self.last_error = exc
+        self.close()
+        self.error_all_callbacks(exc)
+        self.connected_event.set()
+        return exc
+
+    def error_all_callbacks(self, exc):
+        with self.lock:
+            callbacks = self._callbacks
+            self._callbacks = {}
+        new_exc = ConnectionShutdown(str(exc))
+        for cb in callbacks.values():
+            try:
+                cb(new_exc)
+            except Exception:
+                log.warn("Ignoring unhandled exception while erroring callbacks for a "
+                         "failed connection (%s) to host %s:",
+                         id(self), self.host, exc_info=True)
+
+    def handle_pushed(self, response):
+        log.debug("Message pushed from server: %r", response)
+        for cb in self._push_watchers.get(response.event_type, []):
+            try:
+                cb(response.event_args)
+            except Exception:
+                log.exception("Pushed event handler errored, ignoring:")
+
+    def send_msg(self, msg, cb, wait_for_id=False):
+        if self.is_defunct:
+            raise ConnectionShutdown("Connection to %s is defunct" % self.host)
+        elif self.is_closed:
+            raise ConnectionShutdown("Connection to %s is closed" % self.host)
+
+        if not wait_for_id:
+            try:
+                request_id = self._id_queue.get_nowait()
+            except Queue.Empty:
+                raise ConnectionBusy(
+                    "Connection to %s is at the max number of requests" % self.host)
+        else:
+            request_id = self._id_queue.get()
+
+        self._callbacks[request_id] = cb
+        self.push(msg.to_string(request_id, self.protocol_version, compression=self.compressor))
+        return request_id
+
+    def wait_for_response(self, msg, timeout=None):
+        return self.wait_for_responses(msg, timeout=timeout)[0]
 
     def wait_for_responses(self, *msgs, **kwargs):
-        raise NotImplementedError()
+        timeout = kwargs.get('timeout')
+        waiter = ResponseWaiter(self, len(msgs))
+
+        # busy wait for sufficient space on the connection
+        messages_sent = 0
+        while True:
+            needed = len(msgs) - messages_sent
+            with self.lock:
+                available = min(needed, MAX_STREAM_PER_CONNECTION - self.in_flight)
+                self.in_flight += available
+
+            for i in range(messages_sent, messages_sent + available):
+                self.send_msg(msgs[i], partial(waiter.got_response, index=i), wait_for_id=True)
+            messages_sent += available
+
+            if messages_sent == len(msgs):
+                break
+            else:
+                if timeout is not None:
+                    timeout -= 0.01
+                    if timeout <= 0.0:
+                        raise OperationTimedOut()
+                time.sleep(0.01)
+
+        try:
+            return waiter.deliver(timeout)
+        except OperationTimedOut:
+            raise
+        except Exception, exc:
+            self.defunct(exc)
+            raise
 
     def register_watcher(self, event_type, callback):
         raise NotImplementedError()