+++ tests/integration/standard/test_cluster.py
@@ -7,7 +7,6 @@ except ImportError:
 
 import cassandra
 from cassandra.query import SimpleStatement, TraceUnavailable
-from cassandra.io.asyncorereactor import AsyncoreConnection
 from cassandra.policies import RoundRobinPolicy, ExponentialReconnectionPolicy, RetryPolicy, SimpleConvictionPolicy, HostDistance
 
 from cassandra.cluster import Cluster, NoHostAvailable
@@ -88,7 +87,6 @@ class ClusterTests(unittest.TestCase):
             reconnection_policy=ExponentialReconnectionPolicy(1.0, 600.0),
             default_retry_policy=RetryPolicy(),
             conviction_policy_factory=SimpleConvictionPolicy,
-            connection_class=AsyncoreConnection,
             protocol_version=PROTOCOL_VERSION
         )
 
@@ -198,10 +196,17 @@ class ClusterTests(unittest.TestCase):
 
         self.assertRaises(TypeError, session.execute, "SELECT * FROM system.local", trace=True)
 
+        def check_trace(trace):
+            self.assertIsNot(None, trace.request_type)
+            self.assertIsNot(None, trace.duration)
+            self.assertIsNot(None, trace.started_at)
+            self.assertIsNot(None, trace.coordinator)
+            self.assertIsNot(None, trace.events)
+
         query = "SELECT * FROM system.local"
         statement = SimpleStatement(query)
         session.execute(statement, trace=True)
-        self.assertEqual(query, statement.trace.parameters['query'])
+        check_trace(statement.trace)
 
         query = "SELECT * FROM system.local"
         statement = SimpleStatement(query)
@@ -211,12 +216,17 @@ class ClusterTests(unittest.TestCase):
         statement2 = SimpleStatement(query)
         future = session.execute_async(statement2, trace=True)
         future.result()
-        self.assertEqual(query, future.get_query_trace().parameters['query'])
+        check_trace(future.get_query_trace())
 
         statement2 = SimpleStatement(query)
         future = session.execute_async(statement2)
         future.result()
         self.assertEqual(None, future.get_query_trace())
+
+        prepared = session.prepare("SELECT * FROM system.local")
+        future = session.execute_async(prepared, parameters=(), trace=True)
+        future.result()
+        check_trace(future.get_query_trace())
 
     def test_trace_timeout(self):
         cluster = Cluster(protocol_version=PROTOCOL_VERSION)